fix(compiler): simplify generated Int32 conversions (#60)
Generated Go wrapped values that were already native int32 expressions in another int32 conversion. This change widens those proven expressions directly to Go int, which removes redundant conversion lint findings without changing Haxe Int32 behavior.\n\nThe shared file runtime now documents why its ordinary user-visible file modes intentionally honor the process umask. A focused contract protects both lint outcomes, and committed snapshots and examples are refreshed from the compiler output.\n\nVerified with the changed and full test suites, 332 generated snapshots, semantic comparison, strict standard-library sweep, all examples, focused Go lint checks, both performance review harnesses, and the repository pre-commit security checks.

+++ test/snapshot/core/array_identity/intended/main.go
@@ -140,7 +140,7 @@ func main() {
 		hx_array_target_37 := base
 		hx_array_index_38 := index
 		var hx_array_current_39 int = hxrt.IntFromNullableAny(hx_array_target_37.Get(hx_array_index_38))
-		var hx_array_assigned_40 int = int(int32((hxrt.Int32Wrap(hx_array_current_39) + hxrt.Int32Wrap(markedValue()))))
+		var hx_array_assigned_40 int = int((hxrt.Int32Wrap(hx_array_current_39) + hxrt.Int32Wrap(markedValue())))
 		hx_array_target_37.Set(hx_array_index_38, hx_array_assigned_40)
 		return hx_array_assigned_40
 	}()
