+++ test/snapshot/core/unsigned_shift_assign/intended/main.go
@@ -5,7 +5,7 @@ import "snapshot/hxrt"
 func main() {
 	a := -1
 	b := func() int {
-		a = int(int32(int32((uint32(hxrt.Int32Wrap(a)) >> uint(1)))))
+		a = int(int32((uint32(hxrt.Int32Wrap(a)) >> uint(1))))
 		return a
 	}()
 	var v any = any(hxrt.StdString(a))
