+++ test/snapshot/stdlib/haxe_enum_helpers_direct/intended/module_haxe_enumflags.go
@@ -3,7 +3,7 @@ package main
 import "snapshot/hxrt"
 
 func haxe___EnumFlags__EnumFlags_Impl__from(e any) int {
-	i := int(int32((hxrt.Int32Wrap(1) << uint(Type_enumIndex(e)))))
+	i := int((hxrt.Int32Wrap(1) << uint(Type_enumIndex(e))))
 	var this1 int
 	this1 = i
 	return this1
